+++ lib/src/service/polls/poll_field.dart
@@ -3,7 +3,7 @@
 // modification, are permitted provided the conditions.
 
 // Project imports:
-import '../serializable.dart';
+import '../common/serializable.dart';
 
 enum PollField implements Serializable {
   /// `id`
