+++ lib/src/service/tweets/tweet_field.dart
@@ -3,7 +3,7 @@
 // modification, are permitted provided the conditions.
 
 // Project imports:
-import '../serializable.dart';
+import '../common/serializable.dart';
 
 enum TweetField implements Serializable {
   /// `attachments`
