+++ lib/twitter_api_v2.dart
@@ -4,18 +4,20 @@
 
 export 'package:twitter_api_v2/src/auth/oauth_utils.dart';
 export 'package:twitter_api_v2/src/client/oauth_tokens.dart';
+export 'package:twitter_api_v2/src/service/common/includes.dart';
+export 'package:twitter_api_v2/src/service/common/mention.dart';
+export 'package:twitter_api_v2/src/service/common/tag.dart';
+export 'package:twitter_api_v2/src/service/common/url.dart';
 export 'package:twitter_api_v2/src/service/compliance/compliance_data.dart';
 export 'package:twitter_api_v2/src/service/compliance/compliance_service.dart';
 export 'package:twitter_api_v2/src/service/compliance/job_status.dart';
 export 'package:twitter_api_v2/src/service/compliance/job_type.dart';
-export 'package:twitter_api_v2/src/service/includes.dart';
 export 'package:twitter_api_v2/src/service/lists/list_data.dart';
 export 'package:twitter_api_v2/src/service/lists/list_expansion.dart';
 export 'package:twitter_api_v2/src/service/lists/list_meta.dart';
 export 'package:twitter_api_v2/src/service/lists/lists_service.dart';
 export 'package:twitter_api_v2/src/service/media/media_data.dart';
 export 'package:twitter_api_v2/src/service/media/media_type.dart';
-export 'package:twitter_api_v2/src/service/mention.dart';
 export 'package:twitter_api_v2/src/service/places/place_data.dart';
 export 'package:twitter_api_v2/src/service/places/place_field.dart';
 export 'package:twitter_api_v2/src/service/places/place_geo.dart';
@@ -29,7 +31,6 @@ export 'package:twitter_api_v2/src/service/spaces/space_language.dart';
 export 'package:twitter_api_v2/src/service/spaces/space_meta.dart';
 export 'package:twitter_api_v2/src/service/spaces/space_state.dart';
 export 'package:twitter_api_v2/src/service/spaces/spaces_service.dart';
-export 'package:twitter_api_v2/src/service/tag.dart';
 export 'package:twitter_api_v2/src/service/tweets/filtering_rule_data.dart';
 export 'package:twitter_api_v2/src/service/tweets/filtering_rule_meta.dart';
 export 'package:twitter_api_v2/src/service/tweets/geo.dart';
@@ -52,7 +53,6 @@ export 'package:twitter_api_v2/src/service/tweets/tweet_type.dart';
 export 'package:twitter_api_v2/src/service/tweets/tweet_withheld.dart';
 export 'package:twitter_api_v2/src/service/tweets/tweets_service.dart';
 export 'package:twitter_api_v2/src/service/twitter_response.dart';
-export 'package:twitter_api_v2/src/service/url.dart';
 export 'package:twitter_api_v2/src/service/users/user_data.dart';
 export 'package:twitter_api_v2/src/service/users/user_expansion.dart';
 export 'package:twitter_api_v2/src/service/users/user_field.dart';
