+++ test/src/service/twitter_response_test.dart
@@ -6,7 +6,7 @@
 import 'package:test/test.dart';
 
 // Project imports:
-import 'package:twitter_api_v2/src/service/includes.dart';
+import 'package:twitter_api_v2/src/service/common/includes.dart';
 import 'package:twitter_api_v2/src/service/twitter_response.dart';
 
 void main() {
